First 91 numbers of fibonacci series correction using float

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -6,15 +6,15 @@
 
 int main(void)
 {
-	unsigned long long int n, new, prev = 1, ant = 1;
+	float n, new, prev = 1, ant = 1;
 
 	for (n = 1; n < 91; n++)
 	{
-	printf("%llu, ", prev);
+	printf("%.0f, ", prev);
 	new = prev + ant;
 	ant = prev;
 	prev = new;
 	}
-	printf("%llu\n", prev);
+	printf("%.0f\n", prev);
 return (0);
 }
